♻️ refactor(admin): extract shared types and utilities to separate files
- Move ReviewSortField, SortDir types and validation constants to _types.ts
- Extract parseGuessHistory utility function to _utils.ts
- Update imports across admin components to use centralized types
- Add unstable_rethrow for proper error handling in connections action
- Update comment in coderabbit config for clarity
