+++ app/[locale]/admin/_types.ts
@@ -1,0 +1,10 @@
+export type ReviewSortField = "createdAt" | "difficulty" | "reviewerName";
+export type SortDir = "asc" | "desc";
+
+export const VALID_SORT_FIELDS: ReviewSortField[] = [
+  "createdAt",
+  "difficulty",
+  "reviewerName",
+];
+
+export const VALID_SORT_DIRS: SortDir[] = ["asc", "desc"];
